[backend] Migrate injectstatus to InjectExecutions
