+++ openbas-api/src/main/java/io/openbas/execution/ExecutionExecutorService.java
@@ -32,7 +32,9 @@ public class ExecutionExecutorService {
 
   public void launchExecutorContext(Inject inject) {
     InjectExecution injectExecution =
-        inject.getExecutions().orElseThrow(() -> new IllegalArgumentException("Status should exist"));
+        inject
+            .getExecutions()
+            .orElseThrow(() -> new IllegalArgumentException("Status should exist"));
     // First, get the agents and the assets agentless of this inject
     AgentsAndAssetsAgentless agentsAndAssetsAgentless =
         this.injectService.getAgentsAndAgentlessAssetsByInject(inject);
@@ -165,7 +167,8 @@ public class ExecutionExecutorService {
   }
 
   @VisibleForTesting
-  public void saveAgentlessAssetsTraces(Set<Asset> assetsAgentless, InjectExecution injectExecution) {
+  public void saveAgentlessAssetsTraces(
+      Set<Asset> assetsAgentless, InjectExecution injectExecution) {
     if (!assetsAgentless.isEmpty()) {
       executionTraceRepository.saveAll(
           assetsAgentless.stream()
