+++ openbas-api/src/main/java/io/openbas/injectors/lade/service/LadeListener.java
@@ -4,7 +4,7 @@ import io.openbas.database.model.ExecutionStatus;
 import io.openbas.database.model.Inject;
 import io.openbas.database.model.InjectExecution;
 import io.openbas.database.repository.InjectRepository;
-import io.openbas.database.repository.InjectStatusRepository;
+import io.openbas.database.repository.InjectExecutionRepository;
 import io.openbas.injectors.lade.LadeContract;
 import io.openbas.injectors.lade.model.LadeWorkflow;
 import java.time.Instant;
@@ -21,14 +21,14 @@ import org.springframework.stereotype.Service;
 public class LadeListener {
 
   private final InjectRepository injectRepository;
-  private final InjectStatusRepository injectStatusRepository;
+  private final InjectExecutionRepository injectExecutionRepository;
   private final LadeService ladeService;
 
   @Scheduled(fixedDelay = 15000, initialDelay = 0)
   public void listenWorkflows() {
     // Get all lade inject with workflow_id that are not done yet
     List<InjectExecution> injectExecutions =
-        this.injectStatusRepository.pendingForInjectType(LadeContract.TYPE);
+        this.injectExecutionRepository.pendingForInjectType(LadeContract.TYPE);
     // For each workflow ask for traces and status
     injectExecutions.forEach(
         injectStatus -> {
