+++ openbas-api/src/test/java/io/openbas/service/InjectTestStatusServiceTest.java
@@ -239,7 +239,8 @@
 //    InjectTestExecution test = injectTestStatusService.testInject(INJECT1.getId());
 //
 //    // -- EXECUTE --
-//    InjectTestExecution foundTest = injectTestStatusService.findInjectTestStatusById(test.getId());
+//    InjectTestExecution foundTest =
+// injectTestStatusService.findInjectTestStatusById(test.getId());
 //    assertNotNull(foundTest);
 //
 //    // -- CLEAN --
