+++ openbas-model/src/main/java/io/openbas/database/model/Inject.java
@@ -159,23 +159,21 @@ public class Inject implements Base, Injection {
   @JsonProperty("inject_user")
   @Schema(type = "string")
   private User user;
-//
-//  @Getter
-//  @Column(name = "inject_status")
-//  @JsonProperty("inject_status")
-//  @Enumerated(EnumType.STRING)
-//  private ExecutionStatus status;
-//
-//  @Getter
-//  @Column(name = "inject_execution_date")
-//  @JsonProperty("inject_execution_date")
-//  private Instant executionDate;
+
+  //
+  //  @Getter
+  //  @Column(name = "inject_status")
+  //  @JsonProperty("inject_status")
+  //  @Enumerated(EnumType.STRING)
+  //  private ExecutionStatus status;
+  //
+  //  @Getter
+  //  @Column(name = "inject_execution_date")
+  //  @JsonProperty("inject_execution_date")
+  //  private Instant executionDate;
 
   @Getter
-  @OneToMany(
-      mappedBy = "inject",
-      cascade = CascadeType.ALL,
-      orphanRemoval = true)
+  @OneToMany(mappedBy = "inject", cascade = CascadeType.ALL, orphanRemoval = true)
   @JsonIgnore
   private List<InjectExecution> executions;
