+++ openbas-model/src/main/java/io/openbas/database/repository/InjectExecutionRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InjectStatusRepository
+public interface InjectExecutionRepository
     extends CrudRepository<InjectExecution, String>, JpaSpecificationExecutor<InjectExecution> {
 
   @NotNull
