+++ openbas-model/src/main/java/io/openbas/database/repository/InjectTestStatusRepository.java
@@ -8,7 +8,8 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
 public interface InjectTestStatusRepository
-    extends CrudRepository<InjectTestExecution, String>, JpaSpecificationExecutor<InjectTestExecution> {
+    extends CrudRepository<InjectTestExecution, String>,
+        JpaSpecificationExecutor<InjectTestExecution> {
 
   @NotNull
   Optional<InjectTestExecution> findById(@NotNull String id);
